refactor: kill if statement in card handler
